Employee dto

+++ src/main/java/dk/ek/adventureproject/service/EmployeeService.java
@@ -1,10 +1,14 @@
 package dk.ek.adventureproject.Service;
 
 import dk.ek.adventureproject.Model.Employee;
+import dk.ek.adventureproject.dto.EmployeeCreateDTO;
+import dk.ek.adventureproject.dto.EmployeeDTO;
+import dk.ek.adventureproject.dto.Mapper;
 import dk.ek.adventureproject.repo.EmployeeRepo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,41 +17,35 @@ import java.util.Optional;
 public class EmployeeService {
     private final EmployeeRepo employeeRepo;
 
-
-    // Create
-    public Employee createEmployee(Employee employee) {
-        return employeeRepo.save(employee);
+    //using employeeCREATEdto instead, uses password
+    public EmployeeDTO createEmployee(EmployeeCreateDTO dto) {
+        Employee employee = Mapper.toEntity(dto);
+        return Mapper.toDto(employeeRepo.save(employee));
     }
 
-    // get all
-    public List<Employee> getAllEmployees() {
-        return employeeRepo.findAll();
-    }
-
-    // get by id
-    public Optional<Employee> getEmployeeById(Long id) {
-        return employeeRepo.findById(id);
-    }
-
-    // Update
-    public Employee updateEmployee(Long id, Employee employeeDetails) {
-        Optional<Employee> optionalEmployee = employeeRepo.findById(id);
-
-        if (optionalEmployee.isPresent()) {
-            Employee employee = optionalEmployee.get();
-            employee.setUserName(employeeDetails.getUserName());
-            employee.setPassword(employeeDetails.getPassword());
-            employee.setRole(employeeDetails.getRole());
-
-            return employeeRepo.save(employee);
-        } else {
-            throw new RuntimeException("Employee not found with id " + id);
+    public List<EmployeeDTO> getAllEmployees() {
+        List<Employee> employees = employeeRepo.findAll();
+        List<EmployeeDTO> result = new ArrayList<>();
+        for (Employee e : employees) {
+            result.add(Mapper.toDto(e));
         }
+        return result;
     }
 
-    // Delete
+    public EmployeeDTO getEmployeeById(Long id) {
+        Optional<Employee> employee = employeeRepo.findById(id);
+        if (employee.isPresent()) {
+            return Mapper.toDto(employee.get());
+        }
+        throw new RuntimeException("Employee not found");
+    }
+
     public void deleteEmployee(Long id) {
-        employeeRepo.deleteById(id);
+        if(employeeRepo.existsById(id)) {
+            employeeRepo.deleteById(id);
+        } else {
+            throw new RuntimeException("Employee not found");
+        }
     }
 
 }
